semantic-segmentation/requirement.txt
snet.sdk
snet-cli